requirements.txt
allure-pytest
Appium-Python-Client
dotenv
pytest
pytest-xdist
playwright
behave
pytest-playwright
allure-pytest
openpyxl
python-dotenv